aider: Add CMD directive to run "aider" on container start.
# Aider chat conversation:

USER: ls

ASSISTANT: I'm sorry for any confusion, but as an AI, I don't have the ability to execute commands or access file systems. I can only provide guidance on how to modify code based on the information you provide. If you need help with a specific code modification, please provide the details and I'll do my best to assist you.

USER: modify the dockerfile to run the command "aider" on container start

ASSISTANT: To modify the Dockerfile to run the command "aider" on container start, we need to add an ENTRYPOINT or CMD directive at the end of the Dockerfile.

Here is the proposed change:

```dockerfile
agent_harness/aider_config/dockerfile
<<<<<<< HEAD
# Define the directory as a mount point
VOLUME ["/code"]
=======
# Define the directory as a mount point
VOLUME ["/code"]

# Run the command "aider" on container start
CMD ["aider"]
>>>>>>> updated
```

This change will make Docker run the command "aider" every time a container is started from the image. Please note that the "aider" command should be available in the PATH for the Docker container to be able to execute it.